+++ packages/dart/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `createDirectory` method
 - Add `getParentFilePath` method
+- Add `deleteFile` method
 
 ## 1.1.7 (2025-11-03)